Move obfuscation functions to lib/transport/ntcp
